+++ scripts/tests/calibnet_export_check.sh
@@ -18,9 +18,9 @@ echo "Cleaning up the initial snapshot"
 rm --force --verbose ./*.{car,car.zst,sha256sum}
 
 output=$($FOREST_CLI_PATH snapshot export-status --format json)
-is_exporting=$(echo "$output" | jq -r '.exporting')
+state=$(echo "$output" | jq -r '.state')
 echo "Testing that no export is in progress"
-if [ "$is_exporting" == "true" ]; then
+if [ "$state" == "Running" ]; then
   exit 1
 fi
 
@@ -29,8 +29,8 @@ $FOREST_CLI_PATH snapshot export --format "$format" > snapshot_export.log 2>&1 &
 echo "Testing that export is in progress"
 for ((i=1; i<=retries; i++)); do
     output=$($FOREST_CLI_PATH snapshot export-status --format json)
-    is_exporting=$(echo "$output" | jq -r '.exporting')
-    if [ "$is_exporting" == "true" ]; then
+    state=$(echo "$output" | jq -r '.state')
+    if [ "$state" == "Running" ]; then
         break
     fi
     if [ $i -eq $retries ]; then
@@ -45,9 +45,8 @@ $FOREST_CLI_PATH snapshot export-cancel
 echo "Testing that export has been cancelled"
 for ((i=1; i<=retries; i++)); do
     output=$($FOREST_CLI_PATH snapshot export-status --format json)
-    is_exporting=$(echo "$output" | jq -r '.exporting')
-    is_cancelled=$(echo "$output" | jq -r '.cancelled')
-    if [ "$is_exporting" == "false" ] && [ "$is_cancelled" == "true" ]; then
+    state=$(echo "$output" | jq -r '.state')
+    if [ "$state" == "Cancelled" ]; then
         break
     fi
     if [ $i -eq $retries ]; then
@@ -66,7 +65,7 @@ EXPORT_CMD_PID=$!
 sleep 5
 # another export job should be disallowed
 export_error=$($FOREST_CLI_PATH snapshot export 2>&1 || true)
-if echo "$export_error" | grep -q "active chain export job has started"; then
+if echo "$export_error" | grep -q "export has been running since"; then
     echo "verified another export job is disallowed"
 else 
     echo "another export job should be disallowed"
@@ -75,7 +74,7 @@ else
 fi
 # another export-diff job should be disallowed
 export_diff_error=$($FOREST_CLI_PATH snapshot export-diff --from 11000 --to 10100 -d 900 2>&1 || true)
-if echo "$export_diff_error" | grep -q "active chain export job has started"; then
+if echo "$export_diff_error" | grep -q "export has been running since"; then
     echo "verified another export-diff job is disallowed"
 else 
     echo "another export-diff job should be disallowed"